\name{tops}
\alias{tops-package}
\alias{ToPS}
\docType{package}
\title{
A R package for accessing ToPS 
}
\description{
ToPS is an object oriented framework written in C++ that provides a set of probabilistic model to analyze sequence data.
}
\details{
\tabular{ll}{
Package: \tab tops\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-04-13\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}
}
\author{
Maintainer: André Yoshiaki Kashiwabara <akashiwabara@usp.br>
}
\keyword{ package }
\seealso{
  \code{\link{tops_load}}  \code{\link{tops_simulate}}
  \code{\link{tops_evaluate}}  \code{\link{tops_train}}
  \code{\link{tops_viterbi}}   \code{\link{tops_save}}.
}
